+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Fixed
+
+- `sort_scopes()` dropping overlay/local scopes when re-running `setup.sh`
+
 ## [1.0.0] - 2026-04-25
 
 ### Added